requirements.txt
# 知識圖譜構建和可視化
networkx==3.0
pyvis==0.3.2

# LLM和語言模型
langchain==0.1.14
langchain-core==0.1.30
langchain-openai==0.1.3

# 數據處理
matplotlib==3.8.2
pandas==2.1.3
numpy==1.24.3

# 工具和實用程序
requests==2.31.0
python-dotenv==1.0.0
